moving tests, removing compounds to components
