--- src/konnektor/network_planners/concatenator/__init__.py
@@ -1,1 +1,0 @@
-from .mst_concatenator import MstConcatenate
