+++ src/konnektor/network_planners/concatenators/__init__.py
@@ -1,0 +1,1 @@
+from .mst_concatenator import MstConcatenator
