+++ src/konnektor/network_planners/generators/two_dimensional_network_planners.py
@@ -14,7 +14,7 @@ from ._abstract_network_generator import NetworkGenerator
 
 from .cyclic_network_planner import CyclicNetworkGenerator
 from .star_network_planner import StarNetworkGenerator
-from ..concatenator import MstConcatenate
+from ..concatenators import MstConcatenator
 from ...network_tools import append_node, concatenate_networks
 from konnektor.network_tools.clustering.component_diversity_clustering import ComponentsDiversityClusterer
 
@@ -26,7 +26,7 @@ log.setLevel(logging.INFO)
 class TwoDimensionalNetworkGenerator(NetworkGenerator):
     def __init__(self,
                  sub_network_planners: Iterable[NetworkGenerator] = (CyclicNetworkGenerator,),
-                 concatenator: MstConcatenate = MstConcatenate,
+                 concatenator: MstConcatenator = MstConcatenator,
                  clusterer: ComponentsDiversityClusterer = ComponentsDiversityClusterer(
                      featurize=RDKitFingerprintTransformer(), cluster=KMeans(n_clusters=3)),
                  mapper: AtomMapper = None, scorer = None,
@@ -40,12 +40,12 @@ class TwoDimensionalNetworkGenerator(NetworkGenerator):
             This class is seperating the Components along the first dimension.
         sub_network_planners: Iterable[NetworkGenerator]
             The clusters, are then seperatley translated to sub networks by the sub_network_planners
-        concatenator: MstConcatenate
-            The concatenator is connecting the different sub networks.
+        concatenator: MstConcatenator
+            The concatenators is connecting the different sub networks.
         mapper: AtomMapper
-            the atom mapper is required, to define the connection between two ligands, if only concatenator or ligandPlanner classes are passed
+            the atom mapper is required, to define the connection between two ligands, if only concatenators or ligandPlanner classes are passed
         scorer: AtomMappingScorer
-            scoring function evaluating an atom mapping, and giving a score between [0,1], if only concatenator or ligandPlanner classes are passed
+            scoring function evaluating an atom mapping, and giving a score between [0,1], if only concatenators or ligandPlanner classes are passed
         progress: bool, optional
             if true a progress bar will be displayed. (default: False)
         nprocesses: int
@@ -163,9 +163,9 @@ class StarrySkyNetworkGenerator(TwoDimensionalNetworkGenerator):
         clusterer: ComponentsDiversityClusterer
             This class is seperating the Components along the first dimension.
         mapper: AtomMapper
-            the atom mapper is required, to define the connection between two ligands, if only concatenator or ligandPlanner classes are passed
+            the atom mapper is required, to define the connection between two ligands, if only concatenators or ligandPlanner classes are passed
         scorer: AtomMappingScorer
-            scoring function evaluating an atom mapping, and giving a score between [0,1], if only concatenator or ligandPlanner classes are passed
+            scoring function evaluating an atom mapping, and giving a score between [0,1], if only concatenators or ligandPlanner classes are passed
         progress: bool, optional
             if true a progress bar will be displayed. (default: False)
         nprocesses: int
@@ -175,7 +175,7 @@ class StarrySkyNetworkGenerator(TwoDimensionalNetworkGenerator):
 
         super().__init__(clusterer=clusterer,
                          sub_network_planners=[StarNetworkGenerator],
-                         concatenator = MstConcatenate,
+                         concatenator = MstConcatenator,
                          mapper=mapper, scorer=scorer,
                          progress=progress,
                          nprocesses=nprocesses,
