+++ src/konnektor/tests/network_planners/concatenators/test_bipartite_max_concatenator.py
@@ -2,19 +2,16 @@
 # For details, see https://github.com/OpenFreeEnergy/konnektor
 
 import pytest
-import numpy as np
-import networkx as nx
 
 from gufe import LigandNetwork
-from konnektor.network_planners.concatenator import MstConcatenate
+from konnektor.network_planners.concatenators.max_concatenator import MaxConcatenator
 from konnektor.tests.network_planners.conf import (GenAtomMapper, genScorer,
                                                    atom_mapping_basic_test_files, ligand_network_ab)
-from konnektor.network_analysis import get_is_connected
 
 
 #more test here also for the params
-def test_mst_network_concatenation(ligand_network_ab):
-    concatenator = MstConcatenate(mapper=GenAtomMapper(), scorer=genScorer)
+def test_max_network_concatenation(ligand_network_ab):
+    concatenator = MaxConcatenator(mapper=GenAtomMapper(), scorer=genScorer)
 
     ln_a, ln_b = ligand_network_ab
     nA = len(ln_a.nodes)
@@ -25,6 +22,5 @@ def test_mst_network_concatenation(ligand_network_ab):
     cn = concatenator.concatenate_networks([ln_a, ln_b])
 
     assert isinstance(cn, LigandNetwork)
-    assert len(cn.nodes) == nA+nB
-    assert len(cn.edges) == eA + eB + concatenator.n_connecting_edges
-    assert get_is_connected(cn)
+    assert len(cn.nodes) == nA + nB
+    assert len(cn.edges) == eA + eB + nA * nB
