--- src/konnektor/tests/network_planners/test_abstract_network_planner.py
@@ -1,19 +1,0 @@
-# This code is part of OpenFE and is licensed under the MIT license.
-# For details, see https://github.com/OpenFreeEnergy/konnektor
-from rdkit import Chem
-import pytest
-import networkx as nx
-
-from rdkit.Chem import AllChem
-
-import konnektor
-from konnektor import network_planners
-import os
-import importlib
-import pytest
-from importlib import resources
-from rdkit import Chem
-from rdkit.Chem import AllChem
-
-import gufe
-from gufe import SmallMoleculeComponent, LigandAtomMapping
